adding end scene dynamic text fields

+++ Assets/Scripts/PlantSlider.cs
@@ -5,7 +5,7 @@ public class PlantSlider : MonoBehaviour
     // Controller
     public Controller controller;
 
-    public void OnSliderChanged(float value)
+    public void OnSliderChanged(int value)
     {
         // Set plant density to value
         controller.numPlants = value;
